win build
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@64831 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficePPTFile/PPTFormatLib/Reader/SlideInfo.h
@@ -100,7 +100,7 @@ public:
 		if (lIndex > nCount)
 			return lIndex;
 
-        for (size_t nIndex = 0; nIndex < std::min(lIndex, nCount); ++nIndex)
+        for (size_t nIndex = 0; nIndex < (std::min)(lIndex, (DWORD)nCount); ++nIndex)
 		{
 			if ((*m_parEmptyPictures)[nIndex])
 				++lResult;
